solve case 3

+++ libs/demo/src/models/VariableStore.ts
@@ -12,7 +12,6 @@ export type Store = {
 type VariableStoreI = {
   add(storeVariable: StoreVariable): void;
   remove(nameIdentifier: StoreVariable['name']): void;
-  isEmpty(): boolean;
   getVariables(): StoreVariable[];
 };
 
@@ -20,18 +19,26 @@ export default class VariableStore implements VariableStoreI {
   private _variables: StoreVariable[] = [];
 
   public add(storeVariable: Parameters<VariableStoreI['add']>[0]) {
-    this._variables.push(storeVariable);
+    if (!this.exists(storeVariable.name)) {
+      this._variables.push(storeVariable);
+    }
   }
 
   public remove(nameIdentifier: Parameters<VariableStoreI['remove']>[0]) {
-    ArrayUtils.deleteObject(this._variables, nameIdentifier);
+    if (!this.isEmpty()) {
+      ArrayUtils.deleteObject(this._variables, nameIdentifier);
+    }
   }
 
-  public isEmpty() {
+  private isEmpty() {
     return this._variables.length === 0;
   }
 
   public getVariables() {
     return this._variables;
+  }
+
+  private exists(nameIdentifier: StoreVariable['name']) {
+    return !!this._variables.find(({ name }) => name === nameIdentifier);
   }
 }
